Everything appears to be functioning so far. Now for more type
