+++ extract_msg/message_base.py
@@ -1,22 +1,30 @@
 import base64
 import email.utils
 import html
+import json
 import logging
+import os
+import pathlib
 import re
+import subprocess
+import zipfile
 
 import bs4
 import chardet
 import compressed_rtf
 import RTFDE
 
+from email.parser import Parser as EmailParser
+from html import escape as htmlEscape
+from typing import Tuple
+
 from . import constants
 from .enums import DeencapType, RecipientType
-from .exceptions import DeencapMalformedData, DeencapNotEncapsulated
+from .exceptions import DataNotFoundError, DeencapMalformedData, DeencapNotEncapsulated, IncompatibleOptionsError, WKError
 from .msg import MSGFile
 from .recipient import Recipient
-from .utils import inputToString, prepareFilename
-from email.parser import Parser as EmailParser
-from typing import Tuple
+from .utils import addNumToDir, addNumToZipDir, createZipOpen, findWk, inputToBytes, inputToString, isEncapsulatedRtf, prepareFilename, rtfSanitizeHtml, rtfSanitizePlain, validateHtml
+from imapclient.imapclient import decode_utf7
 
 
 logger = logging.getLogger(__name__)
@@ -73,8 +81,8 @@ class MessageBase(MSGFile):
         self.__ignoreRtfDeErrors = kwargs.get('ignoreRtfDeErrors', False)
         self.__deencap = kwargs.get('deencapsulationFunc')
         # Initialize properties in the order that is least likely to cause bugs.
-        # TODO have each function check for initialization of needed data so these
-        # lines will be unnecessary.
+        # TODO have each function check for initialization of needed data so
+        # these lines will be unnecessary.
         self.mainProperties
         self.header
         self.recipients
@@ -109,7 +117,8 @@ class MessageBase(MSGFile):
                 if value:
                     value = value.replace(',', self.__recipientSeparator)
 
-            # If the header had a blank field or didn't have the field, generate it manually.
+            # If the header had a blank field or didn't have the field, generate
+            # it manually.
             if not value:
                 # Check if the header has initialized.
                 if self.headerInit():
@@ -118,12 +127,15 @@ class MessageBase(MSGFile):
                 # Get a list of the recipients of the specified type.
                 foundRecipients = tuple(recipient.formatted for recipient in self.recipients if recipient.type == recipientInt)
 
-                # If we found recipients, join them with the recipient separator and a space.
+                # If we found recipients, join them with the recipient separator
+                # and a space.
                 if len(foundRecipients) > 0:
                     value = (self.__recipientSeparator + ' ').join(foundRecipients)
 
-            # Code to fix the formatting so it's all a single line. This allows the user to format it themself if they want.
-            # This should probably be redone to use re or something, but I can do that later. This shouldn't be a huge problem for now.
+            # Code to fix the formatting so it's all a single line. This allows
+            # the user to format it themself if they want. This should probably
+            # be redone to use re or something, but I can do that later. This
+            # shouldn't be a huge problem for now.
             if value:
                 value = value.replace(' \r\n\t', ' ').replace('\r\n\t ', ' ').replace('\r\n\t', ' ')
                 value = value.replace('\r\n', ' ').replace('\r', ' ').replace('\n', ' ')
@@ -178,6 +190,165 @@ class MessageBase(MSGFile):
             logger.info('No RTF body to deencapsulate from.')
         return None
 
+    def dump(self) -> None:
+        """
+        Prints out a summary of the message.
+        """
+        print('Message')
+        print('Subject:', self.subject)
+        print('Date:', self.date)
+        print('Body:')
+        print(self.body)
+
+    def getJson(self) -> str:
+        """
+        Returns the JSON representation of the Message.
+        """
+        return json.dumps({
+            'from': inputToString(self.sender, 'utf-8'),
+            'to': inputToString(self.to, 'utf-8'),
+            'cc': inputToString(self.cc, 'utf-8'),
+            'bcc': inputToString(self.bcc, 'utf-8'),
+            'subject': inputToString(self.subject, 'utf-8'),
+            'date': inputToString(self.date, 'utf-8'),
+            'body': decode_utf7(self.body),
+        })
+
+    def getSaveBody(self, **kwargs) -> bytes:
+        """
+        Returns the plain text body that will be used in saving based on the
+        arguments.
+
+        :param kwargs: Used to allow kwargs expansion in the save function.
+            Arguments absorbed by this are simply ignored.
+        """
+        # Get the type of line endings.
+        crlf = inputToBytes(self.crlf, 'utf-8')
+
+        outputBytes = b'From: ' + inputToBytes(self.sender, 'utf-8') + crlf
+        outputBytes += b'To: ' + inputToBytes(self.to, 'utf-8') + crlf
+        outputBytes += b'Cc: ' + inputToBytes(self.cc, 'utf-8') + crlf
+        outputBytes += b'Bcc: ' + inputToBytes(self.bcc, 'utf-8') + crlf
+        outputBytes += b'Subject: ' + inputToBytes(self.subject, 'utf-8') + crlf
+        outputBytes += b'Date: ' + inputToBytes(self.date, 'utf-8') + crlf
+        outputBytes += b'-----------------' + crlf + crlf
+        outputBytes += inputToBytes(self.body, 'utf-8')
+
+        return outputBytes
+
+    def getSaveHtmlBody(self, preparedHtml : bool = False, charset : str = 'utf-8', **kwargs) -> bytes:
+        """
+        Returns the HTML body that will be used in saving based on the
+        arguments.
+
+        :param preparedHtml: Whether or not the HTML should be prepared for
+            standalone use (add tags, inject images, etc.).
+        :param charset: If the html is being prepared, the charset to use for
+            the Content-Type meta tag to insert. This exists to ensure that
+            something parsing the html can properly determine the encoding (as
+            not having this tag can cause errors in some programs). Set this to
+            `None` or an empty string to not insert the tag (Default: 'utf-8').
+        :param kwargs: Used to allow kwargs expansion in the save function.
+            Arguments absorbed by this are simply ignored.
+
+        :raises BadHtmlError: if :param preparedHtml: is False and the HTML
+            fails to validate.
+        """
+        if self.htmlBody:
+            # Inject the header into the data.
+            data = self.injectHtmlHeader(prepared = preparedHtml)
+
+            # If we are preparing the HTML, then we should
+            if preparedHtml and charset:
+                bs = bs4.BeautifulSoup(data, features = 'html.parser')
+                if not bs.find('meta', {'http-equiv': 'Content-Type'}):
+                    # Setup the attributes for the tag.
+                    tagAttrs = {
+                        'http-equiv': 'Content-Type',
+                        'content': f'text/html; charset={charset}',
+                    }
+                    # Create the tag.
+                    tag = bs4.Tag(parser = bs, name = 'meta', attrs = tagAttrs, can_be_empty_element = True)
+                    # Add the tag to the head section.
+                    if bs.find('head'):
+                        bs.find('head').insert(0, tag)
+                    else:
+                        # If we are here, the head doesn't exist, so let's add
+                        # it.
+                        if bs.find('html'):
+                            # This should always be true, but I want to be safe.
+                            head = bs4.Tag(parser = bs, name = 'head')
+                            head.insert(0, tag)
+                            bs.find('html').insert(0, head)
+
+                    data = bs.prettify('utf-8')
+
+            return data
+        else:
+            return self.htmlBody
+
+    def getSavePdfBody(self, **kwargs) -> bytes:
+        """
+        Returns the PDF body that will be used in saving based on the arguments.
+
+        :param wkPath: Used to manually specify the path of the wkhtmltopdf
+            executable. If not specified, the function will try to find it.
+            Useful if wkhtmltopdf is not on the path. If :param pdf: is False,
+            this argument is ignored.
+        :param wkOptions: Used to specify additional options to wkhtmltopdf.
+            this must be a list or list-like object composed of strings and
+            bytes.
+        :param kwargs: Used to allow kwargs expansion in the save function.
+            Arguments absorbed by this are simply ignored.
+
+        :raises ExecutableNotFound: The wkhtmltopdf executable could not be
+            found.
+        :raises WKError: Something went wrong in creating the PDF body.
+        """
+        # Immediately try to find the executable.
+        wkPath = findWk(kwargs.get('wkPath'))
+
+        # First thing is first, we need to parse our wkOptions if
+        # they exist.
+        wkOptions = kwargs.get('wkOptions')
+        if wkOptions:
+            try:
+                # Try to convert to a list, whatever it is, and
+                # fail if it is not possible.
+                parsedWkOptions = [*wkOptions]
+            except TypeError:
+                raise TypeError(':param wkOptions: must be an iterable, not {type(wkOptions)}.')
+        else:
+            parsedWkOptions = []
+
+        # Confirm that all of our options we now have are either
+        # strings or bytes.
+        if not all(isinstance(option, (str, bytes)) for option in parsedWkOptions):
+            raise TypeError(':param wkOptions: must be an iterable of strings and bytes.')
+
+        # We call the program to convert the html, but give tell it
+        # the data will go in and come out through stdin and stdout,
+        # respectively. This way we don't have to write temporary
+        # files to the disk. We also ask that it be quiet about it.
+        process = subprocess.Popen([wkPath, *parsedWkOptions, '-', '-'], shell = True, stdin = subprocess.PIPE, stdout = subprocess.PIPE, stderr = subprocess.PIPE)
+        # Give the program the data and wait for the program to
+        # finish.
+        output = process.communicate(self.getSaveHtmlBody(**kwargs))
+        if process.returncode != 0:
+            raise WKError(output[1].decode('utf-8'))
+
+        return output[0]
+
+    def getSaveRtfBody(self, **kwargs) -> bytes:
+        """
+        Returns the RTF body that will be used in saving based on the arguments.
+
+        :param kwargs: Used to allow kwargs expansion in the save function.
+            Arguments absorbed by this are simply ignored.
+        """
+        # Inject the header into the data.
+        return self.injectRtfHeader()
+
     def headerInit(self) -> bool:
         """
         Checks whether the header has been initialized.
@@ -187,6 +358,451 @@ class MessageBase(MSGFile):
             return True
         except AttributeError:
             return False
+
+    def injectHtmlHeader(self, prepared : bool = False) -> bytes:
+        """
+        Returns the HTML body from the MSG file (will check that it has one) with
+        the HTML header injected into it.
+
+        :param prepared: Determines whether to be using the standard HTML (False) or
+            the prepared HTML (True) body (Default: False).
+
+        :raises AttributeError: if the correct HTML body cannot be acquired.
+        :raises BadHtmlError: if :param preparedHtml: is False and the HTML fails to
+            validate.
+        """
+        if not self.htmlBody:
+            raise AttributeError('Cannot inject the HTML header without an HTML body attribute.')
+
+        body = None
+
+        # We don't do this all at once because the prepared body is not cached.
+        if prepared:
+            body = self.htmlBodyPrepared
+
+            # If the body is not valid or not found, raise an AttributeError.
+            if not body:
+                raise AttributeError('Cannot find a prepared HTML body to inject into.')
+        else:
+            body = self.htmlBody
+
+        # Validate the HTML.
+        if not validateHtml(body):
+            # If we are not preparing the HTML body, then raise an
+            # exception.
+            if not prepared:
+                raise BadHtmlError('HTML body failed to pass validation.')
+
+            # If we are here, then we need to do what we can to fix the HTML body.
+            # Unfortunately this gets complicated because of the various ways the
+            # body could be wrong. If only the <body> tag is missing, then we just
+            # need to insert it at the end and be done. If both the <html> and
+            # <body> tag are missing, we determine where to put the body tag (around
+            # everything if there is no <head> tag, otherwise at the end) and then
+            # wrap it all in the <html> tag.
+            parser = bs4.BeautifulSoup(body, features = 'html.parser')
+            if not parser.find('html') and not parser.find('body'):
+                if parser.find('head') or parser.find('footer'):
+                    # Create the parser we will be using for the corrections.
+                    correctedHtml = bs4.BeautifulSoup(b'<html></html>', features = 'html.parser')
+                    htmlTag = correctedHtml.find('html')
+
+                    # Iterate over each of the direct descendents of the parser and
+                    # add each to a new tag if they are not the head or footer.
+                    bodyTag = parser.new_tag('body')
+                    # What we are going to be doing will be causing some of the tags
+                    # to be moved out of the parser, and so the iterator will end up
+                    # pointing to the wrong place after that. To compensate we first
+                    # create a tuple and iterate over that.
+                    for tag in tuple(parser.children):
+                        if tag.name.lower() in ('head', 'footer'):
+                            correctedHtml.append(tag)
+                        else:
+                            bodyTag.append(tag)
+
+                    # All the tags should now be properly in the body, so let's
+                    # insert it.
+                    if correctedHtml.find('head'):
+                        correctedHtml.find('head').insert_after(bodyTag)
+                    elif correctedHtml.find('footer'):
+                        correctedHtml.find('footer').insert_before(bodyTag)
+                    else:
+                        # Neither a head or a body are present, so just append it to
+                        # the main tag.
+                        htmlTag.append(bodyTag)
+                else:
+                    # If there is no <html>, <head>, <footer>, or <body> tag, then
+                    # we just add the tags to the beginning and end of the data and
+                    # move on.
+                    body = b'<html><body>' + body + b'</body></html>'
+            elif parser.find('html'):
+                # Found <html> but not <body>.
+                # Iterate over each of the direct descendents of the parser and
+                # add each to a new tag if they are not the head or footer.
+                bodyTag = parser.new_tag('body')
+                # What we are going to be doing will be causing some of the tags
+                # to be moved out of the parser, and so the iterator will end up
+                # pointing to the wrong place after that. To compensate we first
+                # create a tuple and iterate over that.
+                for tag in tuple(parser.find('html').children):
+                    if tag.name and tag.name.lower() not in ('head', 'footer'):
+                        bodyTag.append(tag)
+
+                # All the tags should now be properly in the body, so let's
+                # insert it.
+                if parser.find('head'):
+                    parser.find('head').insert_after(bodyTag)
+                elif parser.find('footer'):
+                    parser.find('footer').insert_before(bodyTag)
+                else:
+                    parser.find('html').insert(0, bodyTag)
+            else:
+                # Found <body> but not <html>. Just wrap everything in the <html>
+                # tags.
+                body = b'<html>' + body + b'</html>'
+
+        def replace(bodyMarker):
+            """
+            Internal function to replace the body tag with itself plus the header.
+            """
+            # I recently had to change this and how it worked. Now we use a new
+            # property of `MSGFile` that returns a special tuple of tuples to define
+            # how to get all of the properties we are formatting. They are all
+            # processed in the same way, making everything neat. By defining them
+            # in each class, any class can specify a completely different set to be
+            # used.
+            return bodyMarker.group() + self.htmlInjectableHeader.format(
+            **{
+                name: inputToString(htmlEscape(getattr(self, prop)), 'utf-8') if getattr(self, prop) else ''
+                for name, prop in self.headerFormatProperties
+            }).encode('utf-8')
+
+        # Use the previously defined function to inject the HTML header.
+        return constants.RE_HTML_BODY_START.sub(replace, body, 1)
+
+    def injectRtfHeader(self) -> bytes:
+        """
+        Returns the RTF body from this MSG file (will check that it has one)
+        with the RTF header injected into it.
+
+        :raises AttributeError: if the RTF body cannot be acquired.
+        :raises RuntimeError: if all injection attempts fail.
+        """
+        if not self.rtfBody:
+            raise AttributeError('Cannot inject the RTF header without an RTF body attribute.')
+
+        # Try to determine which header to use. Also determines how to sanitize the
+        # rtf.
+        if isEncapsulatedRtf(self.rtfBody):
+            injectableHeader = self.rtfEncapInjectableHeader
+            rtfSanitize = rtfSanitizeHtml
+        else:
+            injectableHeader = self.rtfPlainInjectableHeader
+            rtfSanitize = rtfSanitizePlain
+
+        def replace(bodyMarker):
+            """
+            Internal function to replace the body tag with itself plus the header.
+            """
+            return bodyMarker.group() + injectableHeader.format(
+            **{
+                name: inputToString(rtfSanitize(getattr(self, prop)), 'utf-8') if getattr(self, prop) else ''
+                for name, prop in self.headerFormatProperties
+            }).encode('utf-8')
+
+        # Use the previously defined function to inject the RTF header. We are
+        # trying a few different methods to determine where to place the header.
+        data = constants.RE_RTF_BODY_START.sub(replace, self.rtfBody, 1)
+        # If after any method the data does not match the RTF body, then we have
+        # succeeded.
+        if data != self.rtfBody:
+            logger.debug('Successfully injected RTF header using first method.')
+            return data
+
+        # This second method only applies to encapsulated HTML, so we need to check
+        # for that first.
+        if isEncapsulatedRtf(self.rtfBody):
+            data = constants.RE_RTF_ENC_BODY_START_1.sub(replace, self.rtfBody, 1)
+            if data != self.rtfBody:
+                logger.debug('Successfully injected RTF header using second method.')
+                return data
+
+            # This third method is a lot less reliable, and actually would just
+            # simply violate the encapuslated html, so for this one we don't even
+            # try to worry about what the html will think about it. If it injects,
+            # we swap to basic and then inject again, more worried about it working
+            # than looking nice inside.
+            if constants.RE_RTF_ENC_BODY_UGLY.sub(replace, self.rtfBody, 1) != self.rtfBody:
+                injectableHeader = constants.RTF_PLAIN_INJECTABLE_HEADER
+                data = constants.RE_RTF_ENC_BODY_UGLY.sub(replace, self.rtfBody, 1)
+                logger.debug('Successfully injected RTF header using third method.')
+                return data
+
+        # Severe fallback attempts.
+        data = constants.RE_RTF_BODY_FALLBACK_FS.sub(replace, self.rtfBody, 1)
+        if data != self.rtfBody:
+            logger.debug('Successfully injected RTF header using forth method.')
+            return data
+
+        data = constants.RE_RTF_BODY_FALLBACK_F.sub(replace, self.rtfBody, 1)
+        if data != self.rtfBody:
+            logger.debug('Successfully injected RTF header using fifth method.')
+            return data
+
+        data = constants.RE_RTF_BODY_FALLBACK_PLAIN.sub(replace, self.rtfBody, 1)
+        if data != self.rtfBody:
+            logger.debug('Successfully injected RTF header using sixth method.')
+            return data
+
+        raise RuntimeError('All injection attempts failed. Please report this to the developer.')
+
+    def save(self, **kwargs):
+        """
+        Saves the message body and attachments found in the message.
+
+        The body and attachments are stored in a folder in the current running
+        directory unless :param customPath: has been specified. The name of the
+        folder will be determined by 3 factors.
+           * If :param customFilename: has been set, the value provided for that
+             will be used.
+           * If :param useMsgFilename: has been set, the name of the file used
+             to create the Message instance will be used.
+           * If the file name has not been provided or :param useMsgFilename:
+             has not been set, the name of the folder will be created using the
+             `defaultFolderName` property.
+           * :param maxNameLength: will force all file names to be shortened
+             to fit in the space (with the extension included in the length). If
+             a number is added to the directory that will not be included in the
+             length, so it is recommended to plan for up to 5 characters extra
+             to be a part of the name. Default is 256.
+
+        It should be noted that regardless of the value for maxNameLength, the
+        name of the file containing the body will always have the name 'message'
+        followed by the full extension.
+
+        There are several parameters used to determine how the message will be
+        saved. By default, the message will be saved as plain text. Setting one
+        of the following parameters to True will change that:
+           * :param html: will output the message in HTML format.
+           * :param json: will output the message in JSON format.
+           * :param raw: will output the message in a raw format.
+           * :param rtf: will output the message in RTF format.
+
+        Usage of more than one formatting parameter will raise an exception.
+
+        Using HTML or RTF will raise an exception if they could not be retrieved
+        unless you have :param allowFallback: set to True. Fallback will go in
+        this order, starting at the top most format that is set:
+           * HTML
+           * RTF
+           * Plain text
+
+        If you want to save the contents into a ZipFile or similar object,
+        either pass a path to where you want to create one or pass an instance
+        to :param zip:. If :param zip: is set, :param customPath: will refer to
+        a location inside the zip file.
+
+        :param attachmentsOnly: Turns off saving the body and only saves the
+            attachments when set.
+        :param skipAttachments: Turns off saving attachments.
+        :param charset: If the html is being prepared, the charset to use for
+            the Content-Type meta tag to insert. This exists to ensure that
+            something parsing the html can properly determine the encoding (as
+            not having this tag can cause errors in some programs). Set this to
+            `None` or an empty string to not insert the tag (Default: 'utf-8').
+        :param kwargs: Used to allow kwargs expansion in the save function.
+        :param preparedHtml: When set, prepares the HTML body for standalone
+            usage, doing things like adding tags, injecting attachments, etc.
+            This is useful for things like trying to convert the HTML body
+            directly to PDF.
+        :param pdf: Used to enable saving the body as a PDF file.
+        :param wkPath: Used to manually specify the path of the wkhtmltopdf
+            executable. If not specified, the function will try to find it.
+            Useful if wkhtmltopdf is not on the path. If :param pdf: is False,
+            this argument is ignored.
+        :param wkOptions: Used to specify additional options to wkhtmltopdf.
+            this must be a list or list-like object composed of strings and
+            bytes.
+        """
+        # Move keyword arguments into variables.
+        _json = kwargs.get('json', False)
+        html = kwargs.get('html', False)
+        rtf = kwargs.get('rtf', False)
+        raw = kwargs.get('raw', False)
+        pdf = kwargs.get('pdf', False)
+        allowFallback = kwargs.get('allowFallback', False)
+        _zip = kwargs.get('zip')
+        maxNameLength = kwargs.get('maxNameLength', 256)
+
+        # Variables involved in the save location.
+        customFilename = kwargs.get('customFilename')
+        useMsgFilename = kwargs.get('useMsgFilename', False)
+        #maxPathLength = kwargs.get('maxPathLength', 255)
+
+        # Track if we are only saving the attachments.
+        attachOnly = kwargs.get('attachmentsOnly', False)
+        # Track if we are skipping attachments.
+        skipAttachments = kwargs.get('skipAttachments', False)
+
+        if pdf:
+            kwargs['preparedHtml'] = True
+
+        # ZipFile handling.
+        if _zip:
+            # `raw` and `zip` are incompatible.
+            if raw:
+                raise IncompatibleOptionsError('The options `raw` and `zip` are incompatible.')
+            # If we are doing a zip file, first check that we have been given a path.
+            if isinstance(_zip, (str, pathlib.Path)):
+                # If we have a path then we use the zip file.
+                _zip = zipfile.ZipFile(_zip, 'a', zipfile.ZIP_DEFLATED)
+                kwargs['zip'] = _zip
+                createdZip = True
+            else:
+                createdZip = False
+            # Path needs to be done in a special way if we are in a zip file.
+            path = pathlib.Path(kwargs.get('customPath', ''))
+            # Set the open command to be that of the zip file.
+            _open = createZipOpen(_zip.open)
+            # Zip files use w for writing in binary.
+            mode = 'w'
+        else:
+            path = pathlib.Path(kwargs.get('customPath', '.')).absolute()
+            mode = 'wb'
+            _open = open
+
+        # Reset this for sub save calls.
+        kwargs['customFilename'] = None
+
+        # Check if incompatible options have been provided in any way.
+        if _json + html + rtf + raw + attachOnly + pdf > 1:
+            raise IncompatibleOptionsError('Only one of the following options may be used at a time: json, raw, html, rtf, attachmentsOnly, pdf.')
+
+        # TODO: insert code here that will handle checking all of the msg files to see if the path with overflow.
+
+        if customFilename:
+            # First we need to validate it. If there are invalid characters, this will detect it.
+            if constants.RE_INVALID_FILENAME_CHARACTERS.search(customFilename):
+                raise ValueError('Invalid character found in customFilename. Must not contain any of the following characters: \\/:*?"<>|')
+            path /= customFilename[:maxNameLength]
+        elif useMsgFilename:
+            if not self.filename:
+                raise ValueError(':param useMsgFilename: is only available if you are using an msg file on the disk or have provided a filename.')
+            # Get the actual name of the file.
+            filename = os.path.split(self.filename)[1]
+            # Remove the extensions.
+            filename = os.path.splitext(filename)[0]
+            # Prepare the filename by removing any special characters.
+            filename = prepareFilename(filename)
+            # Shorted the filename.
+            filename = filename[:maxNameLength]
+            # Check to make sure we actually have a filename to use.
+            if not filename:
+                raise ValueError(f'Invalid filename found in self.filename: "{self.filename}"')
+
+            # Add the file name to the path.
+            path /= filename[:maxNameLength]
+        else:
+            path /= self.defaultFolderName[:maxNameLength]
+
+        # Create the folders.
+        if not _zip:
+            try:
+                os.makedirs(path)
+            except Exception:
+                newDirName = addNumToDir(path)
+                if newDirName:
+                    path = newDirName
+                else:
+                    raise Exception(f'Failed to create directory "{path}". Does it already exist?')
+        else:
+            # In my testing I ended up with multiple files in a zip at the same
+            # location so let's try to handle that.
+            pathCompare = str(path).rstrip('/') + '/'
+            if any(x.startswith(pathCompare) for x in _zip.namelist()):
+                newDirName = addNumToZipDir(path, _zip)
+                if newDirName:
+                    path = newDirName
+                else:
+                    raise Exception(f'Failed to create directory "{path}". Does it already exist?')
+
+        # Update the kwargs.
+        kwargs['customPath'] = path
+
+        if raw:
+            self.saveRaw(path)
+            return self
+
+        # If the user has requested the headers for this file, save it now.
+        if kwargs.get('saveHeader', False):
+            headerText = self._getStringStream('__substg1.0_007D')
+            if not headerText:
+                headerText = constants.HEADER_FORMAT.format(subject = self.subject, **self.header)
+
+            with _open(str(path / 'header.txt'), mode) as f:
+                f.write(headerText.encode('utf-8'))
+
+        try:
+            if not attachOnly:
+                # Check what to save the body with.
+                fext = 'json' if _json else 'txt'
+
+                useHtml = False
+                usePdf = False
+                useRtf = False
+                if html:
+                    if self.htmlBody:
+                        useHtml = True
+                        fext = 'html'
+                    elif not allowFallback:
+                        raise DataNotFoundError('Could not find the htmlBody.')
+
+                if pdf:
+                    if self.htmlBody:
+                        usePdf = True
+                        fext = 'pdf'
+                    elif not allowFallback:
+                        raise DataNotFoundError('Count not find the htmlBody to convert to pdf.')
+
+                if rtf or (html and not useHtml) or (pdf and not usePdf):
+                    if self.rtfBody:
+                        useRtf = True
+                        fext = 'rtf'
+                    elif not allowFallback:
+                        raise DataNotFoundError('Could not find the rtfBody.')
+
+            if not skipAttachments:
+                # Save the attachments.
+                attachmentNames = [attachment.save(**kwargs) for attachment in self.attachments]
+
+            if not attachOnly:
+                with _open(str(path / ('message.' + fext)), mode) as f:
+                    if _json:
+                        emailObj = json.loads(self.getJson())
+                        if not skipAttachments:
+                            emailObj['attachments'] = attachmentNames
+
+                        f.write(inputToBytes(json.dumps(emailObj), 'utf-8'))
+                    elif useHtml:
+                        f.write(self.getSaveHtmlBody(**kwargs))
+                    elif usePdf:
+                        f.write(self.getSavePdfBody(**kwargs))
+                    elif useRtf:
+                        f.write(self.getSaveRtfBody(**kwargs))
+                    else:
+                        f.write(self.getSaveBody(**kwargs))
+
+        except Exception:
+            if not _zip:
+                self.saveRaw(path)
+            raise
+        finally:
+            # Close the ZipFile if this function created it.
+            if _zip and createdZip:
+                _zip.close()
+
+        # Return the instance so that functions can easily be chained.
+        return self
 
     @property
     def bcc(self):
