+++ extract_msg/post.py
@@ -1,10 +1,11 @@
-from . import constants
-from .message import Message
 from typing import Tuple
 
+from . import constants
+from .message_base import MessageBase
+from .utils import inputToBytes, inputToString
 
-# Post is effectively identical to Message, just with a different header.
-class Post(Message):
+
+class Post(MessageBase):
     """
     Class for parsing Post messages.
     """
@@ -14,6 +15,38 @@ class Post(Message):
         # We don't need any special data, just need to override the save
         # functions for getting the body. Past that, we just let Message do all
         # the work.
+
+    def getJson(self) -> str:
+        """
+        Returns the JSON representation of the Message.
+        """
+        return json.dumps({
+            'from': inputToString(self.sender, 'utf-8'),
+            'bcc': inputToString(self.bcc, 'utf-8'),
+            'subject': inputToString(self.subject, 'utf-8'),
+            'date': inputToString(self.date, 'utf-8'),
+            'body': decode_utf7(self.body),
+        })
+
+    def getSaveBody(self, **kwargs) -> bytes:
+        """
+        Returns the plain text body that will be used in saving based on the
+        arguments.
+
+        :param kwargs: Used to allow kwargs expansion in the save function.
+            Arguments absorbed by this are simply ignored.
+        """
+        # Get the type of line endings.
+        crlf = inputToBytes(self.crlf, 'utf-8')
+
+        outputBytes = b'From: ' + inputToBytes(self.sender, 'utf-8') + crlf
+        outputBytes += b'Posted At: ' + inputToBytes(self.date, 'utf-8') + crlf
+        outputBytes += b'Conversation: ' + inputToBytes(self.conversation, 'utf-8') + crlf
+        outputBytes += b'Subject: ' + inputToBytes(self.subject, 'utf-8') + crlf
+        outputBytes += b'-----------------' + crlf + crlf
+        outputBytes += inputToBytes(self.body, 'utf-8')
+
+        return outputBytes
 
     @property
     def conversation(self) -> str:
