+++ extract_msg/utils.py
@@ -24,7 +24,6 @@ import zipfile
 import bs4
 import tzlocal
 
-from html import escape as htmlEscape
 from typing import Dict, List, Tuple, Union
 
 from . import constants
@@ -349,206 +348,6 @@ def hasLen(obj) -> bool:
     Checks if :param obj: has a __len__ attribute.
     """
     return hasattr(obj, '__len__')
-
-
-def injectHtmlHeader(msgFile, prepared : bool = False) -> bytes:
-    """
-    Returns the HTML body from the MSG file (will check that it has one) with
-    the HTML header injected into it.
-
-    :param prepared: Determines whether to be using the standard HTML (False) or
-        the prepared HTML (True) body (Default: False).
-
-    :raises AttributeError: if the correct HTML body cannot be acquired.
-    :raises BadHtmlError: if :param preparedHtml: is False and the HTML fails to
-        validate.
-    """
-    if not hasattr(msgFile, 'htmlBody') or not msgFile.htmlBody:
-        raise AttributeError('Cannot inject the HTML header without an HTML body attribute.')
-
-    body = None
-
-    # We don't do this all at once because the prepared body is not cached.
-    if prepared:
-        if hasattr(msgFile, 'htmlBodyPrepared'):
-            body = msgFile.htmlBodyPrepared
-
-        # If the body is not valid or not found, raise an AttributeError.
-        if not body:
-            raise AttributeError('Cannot find a prepared HTML body to inject into.')
-    else:
-        body = msgFile.htmlBody
-
-    # Validate the HTML.
-    if not validateHtml(body):
-        # If we are not preparing the HTML body, then raise an
-        # exception.
-        if not prepared:
-            raise BadHtmlError('HTML body failed to pass validation.')
-
-        # If we are here, then we need to do what we can to fix the HTML body.
-        # Unfortunately this gets complicated because of the various ways the
-        # body could be wrong. If only the <body> tag is missing, then we just
-        # need to insert it at the end and be done. If both the <html> and
-        # <body> tag are missing, we determine where to put the body tag (around
-        # everything if there is no <head> tag, otherwise at the end) and then
-        # wrap it all in the <html> tag.
-        parser = bs4.BeautifulSoup(body, features = 'html.parser')
-        if not parser.find('html') and not parser.find('body'):
-            if parser.find('head') or parser.find('footer'):
-                # Create the parser we will be using for the corrections.
-                correctedHtml = bs4.BeautifulSoup(b'<html></html>', features = 'html.parser')
-                htmlTag = correctedHtml.find('html')
-
-                # Iterate over each of the direct descendents of the parser and
-                # add each to a new tag if they are not the head or footer.
-                bodyTag = parser.new_tag('body')
-                # What we are going to be doing will be causing some of the tags
-                # to be moved out of the parser, and so the iterator will end up
-                # pointing to the wrong place after that. To compensate we first
-                # create a tuple and iterate over that.
-                for tag in tuple(parser.children):
-                    if tag.name.lower() in ('head', 'footer'):
-                        correctedHtml.append(tag)
-                    else:
-                        bodyTag.append(tag)
-
-                # All the tags should now be properly in the body, so let's
-                # insert it.
-                if correctedHtml.find('head'):
-                    correctedHtml.find('head').insert_after(bodyTag)
-                elif correctedHtml.find('footer'):
-                    correctedHtml.find('footer').insert_before(bodyTag)
-                else:
-                    # Neither a head or a body are present, so just append it to
-                    # the main tag.
-                    htmlTag.append(bodyTag)
-            else:
-                # If there is no <html>, <head>, <footer>, or <body> tag, then
-                # we just add the tags to the beginning and end of the data and
-                # move on.
-                body = b'<html><body>' + body + b'</body></html>'
-        elif parser.find('html'):
-            # Found <html> but not <body>.
-            # Iterate over each of the direct descendents of the parser and
-            # add each to a new tag if they are not the head or footer.
-            bodyTag = parser.new_tag('body')
-            # What we are going to be doing will be causing some of the tags
-            # to be moved out of the parser, and so the iterator will end up
-            # pointing to the wrong place after that. To compensate we first
-            # create a tuple and iterate over that.
-            for tag in tuple(parser.find('html').children):
-                if tag.name and tag.name.lower() not in ('head', 'footer'):
-                    bodyTag.append(tag)
-
-            # All the tags should now be properly in the body, so let's
-            # insert it.
-            if parser.find('head'):
-                parser.find('head').insert_after(bodyTag)
-            elif parser.find('footer'):
-                parser.find('footer').insert_before(bodyTag)
-            else:
-                parser.find('html').insert(0, bodyTag)
-        else:
-            # Found <body> but not <html>. Just wrap everything in the <html>
-            # tags.
-            body = b'<html>' + body + b'</html>'
-
-    def replace(bodyMarker):
-        """
-        Internal function to replace the body tag with itself plus the header.
-        """
-        # I recently had to change this and how it worked. Now we use a new
-        # property of `MSGFile` that returns a special tuple of tuples to define
-        # how to get all of the properties we are formatting. They are all
-        # processed in the same way, making everything neat. By defining them
-        # in each class, any class can specify a completely different set to be
-        # used.
-        return bodyMarker.group() + msgFile.htmlInjectableHeader.format(
-        **{
-            name: inputToString(htmlEscape(getattr(msgFile, prop)), 'utf-8') if getattr(msgFile, prop) else ''
-            for name, prop in msgFile.headerFormatProperties
-        }).encode('utf-8')
-
-    # Use the previously defined function to inject the HTML header.
-    return constants.RE_HTML_BODY_START.sub(replace, body, 1)
-
-
-def injectRtfHeader(msgFile) -> bytes:
-    """
-    Returns the RTF body from the MSG file (will check that it has one) with the
-    RTF header injected into it.
-
-    :raises AttributeError: if the RTF body cannot be acquired.
-    :raises RuntimeError: if all injection attempts fail.
-    """
-    if not hasattr(msgFile, 'rtfBody') or not msgFile.rtfBody:
-        raise AttributeError('Cannot inject the RTF header without an RTF body attribute.')
-
-    # Try to determine which header to use. Also determines how to sanitize the
-    # rtf.
-    if isEncapsulatedRtf(msgFile.rtfBody):
-        injectableHeader = msgFile.rtfEncapInjectableHeader
-        rtfSanitize = rtfSanitizeHtml
-    else:
-        injectableHeader = msgFile.rtfPlainInjectableHeader
-        rtfSanitize = rtfSanitizePlain
-
-    def replace(bodyMarker):
-        """
-        Internal function to replace the body tag with itself plus the header.
-        """
-        return bodyMarker.group() + injectableHeader.format(
-        **{
-            name: inputToString(rtfSanitize(getattr(msgFile, prop)), 'utf-8') if getattr(msgFile, prop) else ''
-            for name, prop in msgFile.headerFormatProperties
-        }).encode('utf-8')
-
-    # Use the previously defined function to inject the RTF header. We are
-    # trying a few different methods to determine where to place the header.
-    data = constants.RE_RTF_BODY_START.sub(replace, msgFile.rtfBody, 1)
-    # If after any method the data does not match the RTF body, then we have
-    # succeeded.
-    if data != msgFile.rtfBody:
-        logger.debug('Successfully injected RTF header using first method.')
-        return data
-
-    # This second method only applies to encapsulated HTML, so we need to check
-    # for that first.
-    if isEncapsulatedRtf(msgFile.rtfBody):
-        data = constants.RE_RTF_ENC_BODY_START_1.sub(replace, msgFile.rtfBody, 1)
-        if data != msgFile.rtfBody:
-            logger.debug('Successfully injected RTF header using second method.')
-            return data
-
-        # This third method is a lot less reliable, and actually would just
-        # simply violate the encapuslated html, so for this one we don't even
-        # try to worry about what the html will think about it. If it injects,
-        # we swap to basic and then inject again, more worried about it working
-        # than looking nice inside.
-        if constants.RE_RTF_ENC_BODY_UGLY.sub(replace, msgFile.rtfBody, 1) != msgFile.rtfBody:
-            injectableHeader = constants.RTF_PLAIN_INJECTABLE_HEADER
-            data = constants.RE_RTF_ENC_BODY_UGLY.sub(replace, msgFile.rtfBody, 1)
-            logger.debug('Successfully injected RTF header using third method.')
-            return data
-
-    # Severe fallback attempts.
-    data = constants.RE_RTF_BODY_FALLBACK_FS.sub(replace, msgFile.rtfBody, 1)
-    if data != msgFile.rtfBody:
-        logger.debug('Successfully injected RTF header using forth method.')
-        return data
-
-    data = constants.RE_RTF_BODY_FALLBACK_F.sub(replace, msgFile.rtfBody, 1)
-    if data != msgFile.rtfBody:
-        logger.debug('Successfully injected RTF header using fifth method.')
-        return data
-
-    data = constants.RE_RTF_BODY_FALLBACK_PLAIN.sub(replace, msgFile.rtfBody, 1)
-    if data != msgFile.rtfBody:
-        logger.debug('Successfully injected RTF header using sixth method.')
-        return data
-
-    raise RuntimeError('All injection attempts failed. Please report this to the developer.')
 
 
 def inputToBytes(stringInputVar, encoding) -> bytes:
